test(cloud_security): consolidate kac policy precedence tests
Move TestSetKACPolicyPrecedence out of the internal cloudsecurity test
package and into the external cloudsecurity_test package alongside the
acceptance test, reaching the unexported resource through a new
internal/cloud_security/export_test.go.

Closes #245
